.ipynb_checkpoints/requirements-checkpoint.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-bigquery
pandas
numpy
